requirements.txt
flask
pillow
python-docx
gunicorn
reportlab